+++ Plutarch/DataRepr/Internal/HList/Utils.hs
@@ -36,6 +36,13 @@ type family IndexList (n :: Nat) (l :: [k]) :: k where
 -- | Indexing list of labeled pairs by label
 type IndexLabel :: Symbol -> [(Symbol, Type)] -> Type
 type family IndexLabel name as where
+  IndexLabel name '[] =
+    TypeError
+      ( 'Text "Invalid field name `"
+          ':<>: 'Text name
+          ':<>: 'Text "`"
+          ':$$: 'Text "Consider adding it to `pletFields` list"
+      )
   IndexLabel name ('(name, a) ': _) = a
   IndexLabel name (_ ': as) = IndexLabel name as