+++ Plutarch/Script.hs
@@ -1,7 +1,11 @@
-module Plutarch.Script (Script (..), serialiseScript, deserialiseScript) where
+module Plutarch.Script (Script (..), serialiseScript, deserialiseScript, hashScriptWithPrefix) where
 
-import Data.ByteString.Short (ShortByteString)
+import Data.ByteString (ByteString)
+import Data.ByteString qualified as BS
+import Data.ByteString.Short (ShortByteString, fromShort)
+import Data.Word (Word8)
 import GHC.Generics (Generic)
+import PlutusCore.Crypto.Hash qualified as Hash
 import PlutusLedgerApi.Common (serialiseUPLC, uncheckedDeserialiseUPLC)
 import UntypedPlutusCore qualified as UPLC
 
@@ -14,3 +18,8 @@ serialiseScript = serialiseUPLC . unScript
 
 deserialiseScript :: ShortByteString -> Script
 deserialiseScript = Script . uncheckedDeserialiseUPLC
+
+hashScriptWithPrefix :: Word8 -> Script -> ByteString
+hashScriptWithPrefix prefix scr =
+  Hash.blake2b_224 $
+    BS.singleton prefix <> (fromShort . serialiseUPLC . unScript $ scr)